flask-dev/flask-dev/openai_client.py
from __future__ import annotations

import os
from dotenv import load_dotenv
from openai import OpenAI

load_dotenv()

# Centralized client so both streaming and non-streaming paths share it

def get_client() -> OpenAI:
    api_key = os.getenv("OPENAI_API_KEY")
    if not api_key:
        raise RuntimeError("OPENAI_API_KEY not set. Put it in .env")
    return OpenAI(api_key=api_key)