flask-dev/flask-dev/requirements.txt
flask==3.0.3
jinja2>=3.1.0
werkzeug>=3.0.0
python-dotenv>=1.0.1
openai>=1.30.0
requests>=2.31.0
